refactor: wire justfile to solobase-browser export-assets
Replace sql-assets target (cross-repo fetch) and hand-rolled
__BUILD_ID__ sed with a single cargo run --manifest-path
../solobase/Cargo.toml -p solobase-browser --bin export-assets
--app-name gizza-ai --app-title 'Gizza AI' --boot-redirect /.
Delete site/sw.js, site/loader.js, site/bridge.js — all three
come from the framework's parameterized templates now.

Move localLlmChatStream from the deleted site/bridge.js into
site/ai-bridge.js (its natural home alongside the WebLLM
inference code) and update the wasm_bindgen module path in
src/blocks/agent.rs from /site/bridge.js to /site/ai-bridge.js.

Gizza's branded index.html and UI scripts are cp'd AFTER
export-assets so gizza's versions win over framework defaults.
